+++ src/Symfony/Component/Serializer/Attribute/SerializedName.php
@@ -16,22 +16,40 @@ use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 /**
  * @author Fabien Bourigault <bourigaultfabien@gmail.com>
  */
-#[\Attribute(\Attribute::TARGET_METHOD | \Attribute::TARGET_PROPERTY)]
+#[\Attribute(\Attribute::TARGET_METHOD | \Attribute::TARGET_PROPERTY | \Attribute::IS_REPEATABLE)]
 class SerializedName
 {
+    private array $groups;
+
     /**
-     * @param string $serializedName The name of the property as it will be serialized
+     * @param string          $serializedName The name of the property as it will be serialized
+     * @param string|string[] $groups         The groups to use when serializing or deserializing
      */
-    public function __construct(private readonly string $serializedName)
-    {
+    public function __construct(
+        private readonly string $serializedName,
+        string|array $groups = ['*'],
+    ) {
         if ('' === $serializedName) {
             throw new InvalidArgumentException(\sprintf('Parameter given to "%s" must be a non-empty string.', self::class));
+        }
+
+        $this->groups = ((array) $groups) ?: ['*'];
+
+        foreach ($this->groups as $group) {
+            if (!\is_string($group)) {
+                throw new InvalidArgumentException(\sprintf('Parameter "groups" given to "%s" must be a string or an array of strings, "%s" given.', static::class, get_debug_type($group)));
+            }
         }
     }
 
     public function getSerializedName(): string
     {
         return $this->serializedName;
+    }
+
+    public function getGroups(): array
+    {
+        return $this->groups;
     }
 }