+++ src/Symfony/Component/Serializer/Tests/Mapping/Loader/XmlFileLoaderTest.php
@@ -80,6 +80,8 @@ class XmlFileLoaderTest extends TestCase
         $attributesMetadata = $classMetadata->getAttributesMetadata();
         $this->assertEquals('baz', $attributesMetadata['foo']->getSerializedName());
         $this->assertEquals('qux', $attributesMetadata['bar']->getSerializedName());
+        $this->assertSame('duxi', $attributesMetadata['duux']->getSerializedName());
+        $this->assertSame('duxa', $attributesMetadata['duux']->getSerializedName(['a']));
     }
 
     public function testSerializedPath()
@@ -90,6 +92,8 @@ class XmlFileLoaderTest extends TestCase
         $attributesMetadata = $classMetadata->getAttributesMetadata();
         $this->assertEquals('[one][two]', $attributesMetadata['three']->getSerializedPath());
         $this->assertEquals('[three][four]', $attributesMetadata['seven']->getSerializedPath());
+        $this->assertSame('[five][six]', $attributesMetadata['eleven']->getSerializedPath());
+        $this->assertSame('[six][five]', $attributesMetadata['eleven']->getSerializedPath(['a']));
     }
 
     public function testSerializedPathInConstructor()
@@ -99,6 +103,8 @@ class XmlFileLoaderTest extends TestCase
 
         $attributesMetadata = $classMetadata->getAttributesMetadata();
         $this->assertEquals('[one][two]', $attributesMetadata['three']->getSerializedPath());
+        $this->assertSame('[five][six]', $attributesMetadata['eleven']->getSerializedPath());
+        $this->assertSame('[six][five]', $attributesMetadata['eleven']->getSerializedPath(['a']));
     }
 
     public function testLoadDiscriminatorMap()