+++ src/Symfony/Component/Serializer/Tests/Mapping/Loader/YamlFileLoaderTest.php
@@ -97,6 +97,8 @@ class YamlFileLoaderTest extends TestCase
         $attributesMetadata = $classMetadata->getAttributesMetadata();
         $this->assertEquals('baz', $attributesMetadata['foo']->getSerializedName());
         $this->assertEquals('qux', $attributesMetadata['bar']->getSerializedName());
+        $this->assertSame('duxi', $attributesMetadata['duux']->getSerializedName());
+        $this->assertSame('duxa', $attributesMetadata['duux']->getSerializedName(['a']));
     }
 
     public function testSerializedPath()
@@ -107,6 +109,8 @@ class YamlFileLoaderTest extends TestCase
         $attributesMetadata = $classMetadata->getAttributesMetadata();
         $this->assertEquals(new PropertyPath('[one][two]'), $attributesMetadata['three']->getSerializedPath());
         $this->assertEquals(new PropertyPath('[three][four]'), $attributesMetadata['seven']->getSerializedPath());
+        $this->assertSame(new PropertyPath('[five][six]'), $attributesMetadata['eleven']->getSerializedPath());
+        $this->assertSame(new PropertyPath('[six][five]'), $attributesMetadata['eleven']->getSerializedPath(['a']));
     }
 
     public function testSerializedPathInConstructor()
@@ -116,6 +120,8 @@ class YamlFileLoaderTest extends TestCase
 
         $attributesMetadata = $classMetadata->getAttributesMetadata();
         $this->assertEquals(new PropertyPath('[one][two]'), $attributesMetadata['three']->getSerializedPath());
+        $this->assertSame(new PropertyPath('[five][six]'), $attributesMetadata['eleven']->getSerializedPath());
+        $this->assertSame(new PropertyPath('[six][five]'), $attributesMetadata['eleven']->getSerializedPath(['a']));
     }
 
     public function testLoadDiscriminatorMap()